requirements.txt
requests
python-dotenv
pylint
openai
atproto